Minor bug fixes
Refactoring of class/function names

Modification of import statements for hardware.awgs.old_base
