+++ qupulse/hardware/awgs/base.py
@@ -1,97 +1,25 @@
-import warnings
 from abc import ABC, abstractmethod
-from copy import copy
-from typing import TypeVar, Generic, List, Iterable, Dict, Callable
+from typing import Iterable, Optional
+
+from .base_features import BaseFeature, FeatureAble
 
 
-class BaseFeature(ABC):
-    """
-    Base class for features of `FeatureAble`s.
-
-    Features are classes containing functions which are bound dynamically to the target instance of type `FeatureAble`.
-    This ensures that all targets for the same feature are using the same signature for the feature's functions. All
-    public callables of a specific feature will be added to the function dictionary. Those functions (in the `functions`
-    dictionary) will be automatically added to the specific `FeatureAble` that calls `FeatureAble.add_feature`.
-    """
-
-    def __init__(self):
-        super().__init__()
-
-        self._functions = self._read_functions()
-
-    def _read_functions(self) -> Dict[str, Callable]:
-        """
-        Reads the functions of a feature and returns them as a dictionary
-
-        Return:
-            Returns dictionary with all functions of the feature
-        """
-        directory = dir(self)
-        function_list = {}
-        for attr in directory:
-            if callable(getattr(type(self), attr)) and attr[0] != "_":
-                if not (attr in function_list):
-                    function_list[attr] = getattr(self, attr)
-        return function_list
-
-    @property
-    def functions(self) -> Dict[str, Callable]:
-        """Returns a copy of the dictionary with all public functions of the feature"""
-        return copy(self._functions)
-
-
-FeatureType = TypeVar(BaseFeature)
-
-
-class FeatureAble(Generic[FeatureType], ABC):
-    """Base class for all classes that are able to add features"""
-
-    def __init__(self):
-        super().__init__()
-
-        self._features = {}
-
-    @property
-    def features(self) -> Dict[str, Callable]:
-        """Returns the dictionary with all features of a FeatureAble"""
-        return copy(self._features)
-
-    def add_feature(self, feature: FeatureType) -> None:
-        """
-        The method adds all functions of feature to a dictionary with all functions
-
-        Args:
-             feature: A certain feature which functions should be added to the dictionary _features
-        """
-        if not isinstance(feature, FeatureType):
-            raise TypeError("Invalid type for feature")
-
-        for function in feature.function_list:
-            if not hasattr(self, function):
-                setattr(self, function, getattr(feature, function))
-            else:
-                warnings.warning(
-                    f"Ommiting function \"{function}\": Another attribute with this name already exists.")
-
-        self._features[type(feature).__name__] = feature
-
-
-class AWGDeviceFeature(BaseFeature, ABC):
+class BaseAWGFeature(BaseFeature, ABC):
     """Base class for features that are used for `AWGDevice`s"""
     pass
 
 
-class AWGChannelFeature(BaseFeature, ABC):
+class BaseAWGChannelFeature(BaseFeature, ABC):
     """Base class for features that are used for `AWGChannel`s"""
     pass
 
 
-class AWGChannelTupleFeature(BaseFeature, ABC):
+class BaseAWGChannelTupleFeature(BaseFeature, ABC):
     """Base class for features that are used for `AWGChannelTuple`s"""
     pass
 
 
-class BaseAWG(FeatureAble[AWGDeviceFeature], ABC):
+class BaseAWG(FeatureAble[BaseAWGFeature], ABC):
     """Base class for all drivers of all arbitrary waveform generators"""
 
     def __init__(self, name: str):
@@ -105,30 +33,30 @@ class BaseAWG(FeatureAble[AWGDeviceFeature], ABC):
     def __del__(self):
         self.cleanup()
 
-    @abstractmethod
-    def cleanup(self) -> None:
-        """Function for cleaning up the dependencies of the device"""
-        pass
-
     @property
     def name(self) -> str:
         """Returns the name of a Device as a String"""
         return self._name
 
+    @abstractmethod
+    def cleanup(self) -> None:
+        """Function for cleaning up the dependencies of the device"""
+        raise NotImplementedError()
+
     @property
     @abstractmethod
     def channels(self) -> Iterable["BaseAWGChannel"]:
         """Returns a list of all channels of a Device"""
-        return self._channels
+        raise NotImplementedError()
 
     @property
     @abstractmethod
     def channel_tuples(self) -> Iterable["BaseAWGChannelTuple"]:
         """Returns a list of all channel tuples of a list"""
-        pass
+        raise NotImplementedError()
 
 
-class BaseAWGChannelTuple(FeatureAble[AWGChannelTupleFeature], ABC):
+class BaseAWGChannelTuple(FeatureAble[BaseAWGChannelTupleFeature], ABC):
     """Base class for all groups of synchronized channels of an AWG"""
 
     def __init__(self, idn: int):
@@ -141,30 +69,30 @@ class BaseAWGChannelTuple(FeatureAble[AWGChannelTupleFeature], ABC):
         self._idn = idn
 
     @property
-    @abstractmethod
-    def sample_rate(self) -> float:
-        """Returns the sample rate of a channel tuple as a float"""
-        pass
-
-    @property
     def idn(self) -> int:
         """Returns the identification number of a channel tuple"""
         return self._idn
 
     @property
     @abstractmethod
+    def sample_rate(self) -> float:
+        """Returns the sample rate of a channel tuple as a float"""
+        raise NotImplementedError()
+
+    @property
+    @abstractmethod
     def device(self) -> BaseAWG:
         """Returns the device which the channel tuple belong to"""
-        pass
+        raise NotImplementedError()
 
     @property
     @abstractmethod
     def channels(self) -> Iterable["BaseAWGChannel"]:
         """Returns a list of all channels of the channel tuple"""
-        pass
+        raise NotImplementedError()
 
 
-class BaseAWGChannel(FeatureAble[AWGChannelFeature], ABC):
+class BaseAWGChannel(FeatureAble[BaseAWGChannelFeature], ABC):
     """Base class for a single channel of an AWG"""
 
     def __init__(self, idn: int):
@@ -184,19 +112,20 @@ class BaseAWGChannel(FeatureAble[AWGChannelFeature], ABC):
     @abstractmethod
     def device(self) -> BaseAWG:
         """Returns the device which the channel belongs to"""
-        pass
+        raise NotImplementedError()
 
     @property
     @abstractmethod
-    def channel_tuple(self) -> BaseAWGChannelTuple:
+    def channel_tuple(self) -> Optional[BaseAWGChannelTuple]:
         """Returns the channel tuple which a channel belongs to"""
-        pass
+        raise NotImplementedError()
 
     @abstractmethod
-    def _set_channel_tuple(self, channel_tuple: BaseAWGChannelTuple):
+    def _set_channel_tuple(self, channel_tuple) -> None:
         """
         Sets the channel tuple which a channel belongs to
+
         Args:
             channel_tuple: reference to the channel tuple
         """
-        pass
+        raise NotImplementedError()
